requirements.txt
torch
fastapi
uvicorn
numpy
Pillow
torchvision
jinja2
python-multipart
google-generativeai